requirements.txt
Django==4.1.4
django-ckeditor==6.5.1
django-sass==1.1.0
django-sass-processor==1.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
requests==2.28.2
sorl-thumbnail==12.9.0
python-dotenv==0.21.0
gunicorn==20.1.0
django-cleanup==7.0.0
djangorestframework==3.14.0
django-filter==22.1